.NET 6.0 (#615)
* net6.0

* merge upstream and update workflows

* use nuget pack not dotnet pack

* name

* setup nuget separately

* change release name back

* readme

* image before releases

* replace webclient with httpclient

* use multiline

* use line concatenation for a couple places

* no slash

* use &&

* break up steps more

* fix xml writer errors and such

* more fixes

* fix httpclient

* code cleanup

* add back comments

+++ mzLib/Test/TestMsDataFile.cs
@@ -188,9 +188,9 @@ namespace Test
 
             var precursorScan = myMsDataFile.GetOneBasedScan(theScan.OneBasedPrecursorScanNumber.Value);
             theScan.RefineSelectedMzAndIntensity(precursorScan.MassSpectrum);
-            Assert.AreEqual(.32872, theScan.SelectedIonIntensity, 0.01);
-            Assert.AreEqual(693.9892, theScan.SelectedIonMZ, 0.01);
-            Assert.AreEqual(693.655, theScan.SelectedIonMonoisotopicGuessMz, 0.001);
+            Assert.AreEqual(.32872, (double)theScan.SelectedIonIntensity, 0.01);
+            Assert.AreEqual(693.9892, (double)theScan.SelectedIonMZ, 0.01);
+            Assert.AreEqual(693.655, (double)theScan.SelectedIonMonoisotopicGuessMz, 0.001);
 
             Assert.AreNotEqual(0, myMsDataFile.GetOneBasedScan(2).MassSpectrum.FirstX);
